backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.18.4
    # via -r requirements.in
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==5.0.0
    # via -r requirements.in
build==1.4.0
    # via pip-tools
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via
    #   pip-tools
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   uvicorn
cryptography==46.0.5
    # via -r requirements.in
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi[standard]==0.135.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.24
    # via fastapi
fastapi-cloud-cli==0.15.1
    # via fastapi-cli
fastar==0.9.0
    # via fastapi-cloud-cli
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
packaging==26.0
    # via
    #   build
    #   wheel
pip-tools==7.5.3
    # via -r requirements.in
pycparser==3.0
    # via cffi
pydantic[email]==2.12.5
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.1
    # via fastapi
pydantic-settings==2.13.1
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.20.0
    # via rich
pyjwt==2.12.1
    # via -r requirements.in
pymysql==1.1.2
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
rich==14.3.3
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.19.7
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.57.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.48
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.37
    # via -r requirements.in
starlette==0.52.1
    # via fastapi
typer==0.24.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via sentry-sdk
uvicorn[standard]==0.41.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
